Move `LocalFileStore` to WordPressShared
We need it there because it's used by `SharedDataIssueSolver` which
we'll soon need to move to WordPressData.
