+++ Modules/Sources/WordPressShared/FileManager/FileManager+FolderSize.swift
